refactor: update format, utils, validation in be and fe

+++ backend/src/services/groupMember.service.ts
@@ -1,4 +1,4 @@
-import { Types } from "mongoose";
+import mongoose, { Types } from "mongoose";
 import GroupModel from "../models/Group.model";
 import GroupMemberModel from "../models/GroupMember.model";
 import UserModel from "../models/User.model";
@@ -120,47 +120,47 @@ const deleteMember = async (groupId: string, memberId: string) => {
 };
 
 const leaveGroup = async (groupId: string, userId: string) => {
-  // 1. Tìm member đang thực hiện yêu cầu
-  const currentMember = await GroupMemberModel.findOne({
-    groupId: new Types.ObjectId(groupId),
-    userId: new Types.ObjectId(userId),
-  });
+  const session = await mongoose.startSession();
+  session.startTransaction();
 
-  if (!currentMember) {
-    throw new ForbiddenException("Bạn không phải là thành viên của nhóm này");
-  }
+  try {
+    const currentMember = await GroupMemberModel.findOne({
+      groupId,
+      userId,
+    }).session(session);
+    if (!currentMember)
+      throw new ForbiddenException("You are not a member of this group");
 
-  // 2. Nếu là OWNER, kiểm tra xem có ai khác cũng là OWNER không
-  if (currentMember.role === "OWNER") {
-    const otherOwner = await GroupMemberModel.findOne({
-      groupId: new Types.ObjectId(groupId),
-      userId: { $ne: new Types.ObjectId(userId) }, // Không phải là chính mình ($ne: Not Equal)
-      role: "OWNER",
-    });
+    if (currentMember.role === "OWNER") {
+      const otherOwnerCount = await GroupMemberModel.countDocuments({
+        groupId,
+        userId: { $ne: userId },
+        role: "OWNER",
+      }).session(session);
 
-    if (!otherOwner) {
-      throw new ForbiddenException(
-        "Bạn là chủ nhóm duy nhất. Hãy chuyển quyền OWNER cho thành viên khác trước khi rời nhóm!",
-      );
+      if (otherOwnerCount === 0) {
+        throw new ForbiddenException(
+          "You are the only owner. Please transfer ownership to another member before leaving!",
+        );
+      }
     }
+
+    await GroupMemberModel.deleteOne({ groupId, userId }).session(session);
+
+    await session.commitTransaction();
+    return { message: "Left group successfully" };
+  } catch (error) {
+    await session.abortTransaction();
+    throw error;
+  } finally {
+    session.endSession();
   }
-
-  // 3. Nếu vượt qua các bước check trên thì cho phép xóa khỏi nhóm
-  await GroupMemberModel.deleteOne({
-    groupId: new Types.ObjectId(groupId),
-    userId: new Types.ObjectId(userId),
-  });
-
-  return {
-    message: "Rời nhóm thành công",
-    data: { groupId, userId },
-  };
 };
 
 export {
   addMember,
   deleteMember,
   getMembersByGroupId,
-  updateMemberRole,
   leaveGroup,
+  updateMemberRole,
 };
